// web/src/shared/api/generated/model/examItemResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AttemptResult } from './attemptResult';
import type { JsonElement } from './jsonElement';

export interface ExamItemResult {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  position: number | string;
  questionId: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  questionVersion: number | string;
  type: string;
  stem: string;
  body: JsonElement;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  maxScore: number | string;
  savedAnswer: null | JsonElement;
  /** @nullable */
  answerSavedAt: string | null;
  attempt: null | AttemptResult;
  correctAnswer: null | JsonElement;
  /** @nullable */
  explanation: string | null;
}